requirements.txt
Flask-MySQL
Flask
flask-mysqldb
mysql-connector-python